requirements.txt
streamlit~=1.29.0
openai~=1.6.1
pytesseract~=0.3.10
pdf2image~=1.17.0

requests~=2.31.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
boto3~=1.34.32
python-dotenv~=0.21.0
matplotlib~=3.8.0
numpy~=1.26.3
pandas~=2.1.4
plotly~=5.9.0